skip_models to set, and renamed testing dir
